First commit, open source buildscripts.
Summary:
Open source our buildbot.py buildscripts for various projects.

This is in a new repo for now, as our buildenv is very specific to us, to the
point of harcoding several paths.

This is something we should fix.

Test Plan: Tested something similar to the `cp -r` we run in meta-town, locally.

Reviewers: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D86